refactor: subtitle service 로직 리팩토링
- service 로직 리팩토링 및 주석
