--- backend/src/subtitles/interfaces/subtitle.interface.ts
@@ -1,10 +1,0 @@
-export interface Subtitle {
-    startTime: string;
-    endTime: string;
-    text: string;
-}
-
-export interface SubtitleResponse {
-    subtitles: Subtitle[];
-    videoId: string;
-}
