+++ backend/src/subtitles/subtitle.interface.ts
@@ -3,3 +3,8 @@ export interface Subtitle {
     endTime: string;
     text: string;
 }
+
+export interface SubtitleResponse {
+    subtitles: Subtitle[];
+    videoId: string;
+}
